Fix similarityScore so that it finds optimal score
Previously I thought the input was the optimal alignment, but it
turns out that similarityScore is also supposed to find that optimal
alignment, or at least the score of it. The implementation now is
the brute force one. Later I will make a memoized version that will
hopefully work much faster.